+++ src/ProjectManager/ProjectManager.ts
@@ -206,6 +206,7 @@ export class ProjectManager implements ProjectManagerInterface {
         projectTree.addIncluded(getFilePathInWorkspace((await getRootProjectFilePath()) || ""));
         projectTree.addIncluded(getFilePathInWorkspace(".vscode"));
         projectTree.addIncluded(getLogPath());
+        projectTree.addIncluded(getFilePathInWorkspace("Package.swift"));
 
         // add all project first as they are visible
         for (const file of await this.projectCache.allFiles()) {